+++ app/src/main/java/cord/eoeo/momentwo/ui/photolist/PhotoListViewModel.kt
@@ -7,7 +7,7 @@ import androidx.navigation.toRoute
 import cord.eoeo.momentwo.core.data.photo.PhotoRepository
 import cord.eoeo.momentwo.core.domain.subalbum.ChangeSubAlbumTitleUseCase
 import cord.eoeo.momentwo.core.common.BaseViewModel
-import cord.eoeo.momentwo.ui.MomentwoDestination
+import cord.eoeo.momentwo.core.navigation.MomentwoDestination
 import cord.eoeo.momentwo.ui.photolist.PhotoListContract.Effect.ShowSnackbar
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
